Refactor suggested changes for direction
Changed Orientation to Direction
Moved Door class into separate kotlin file
Cleanup syntax
